+++ include/emhit.h
@@ -56,7 +56,7 @@ void YarareInitCube(cEm* em, f32 x, f32 y, f32 z, f32 w, f32 h, f32 extent, s16 
 void YarareAdd(cEm* em, YARARE_INFO* box, f32 x, f32 y, f32 z, f32 w, f32 h, s16 no, u16 flags);
 void YarareAddCube(cEm* em, YARARE_INFO* box, f32 x, f32 y, f32 z, f32 w, f32 h, f32 extent, s16 no, u16 flags);
 int EmGetDmPos(cEm* em, Vec* pos, Vec* dir);                                     // em_sub.cpp
-void EmDmBloodSet2(cEm* em, int est_id, int type, int mode, int esp_core_flg, int core_kind);               // em_sub.cpp
+void EmDmBloodSet2(cEm* em, u32 est_id, u32 type, u32 mode, u16 esp_core_flg, u32 core_kind);               // em_sub.cpp
 int VehicleAdjust(Vec* pos);                                                     // em_sub.cpp: rides `pos` along the trolley (room 21B)
 }
 
@@ -71,7 +71,7 @@ struct EmAtkInfo {
 };
 
 extern "C" {
-void EmPlBloodSet2(cModel* m, Vec* pos, int a, int eff_id, int type);                 // em_sub.cpp
+void EmPlBloodSet2(cModel* m, Vec* pos, u32 a, u8 eff_id, u8 type);                 // em_sub.cpp
 // Line `a`-`b` against the enemies: the hit enemy or NULL; hit point / normal and the scenario attribute out.
 cEm* EmAtkLineHitCk(Vec* pPos, Vec* pPos2, Vec* hit, Vec* nrm, u32* attr);              // em_sub.cpp
 void EmAtkSetDamagePL(cEm* em, EmAtkInfo* info, Vec* pPos, Vec* pPos2);                 // em_sub.cpp
